refactor db functions to one file

+++ models/stations.go
@@ -89,7 +89,7 @@ type GetStationResponseSchema struct {
 	Functions            []Function         `json:"functions" bson:"functions"`
 	IsDeleted            bool               `json:"is_deleted" bson:"is_deleted"`
 	Tags                 []CreateTag        `json:"tags"`
-	IdempotencyWindow    int                `json:"idempotency_window_in_ms" bson:"idempotency_window_in_ms"`
+	IdempotencyWindow    int64              `json:"idempotency_window_in_ms" bson:"idempotency_window_in_ms"`
 	IsNative             bool               `json:"is_native" bson:"is_native"`
 	DlsConfiguration     DlsConfiguration   `json:"dls_configuration" bson:"dls_configuration"`
 	TieredStorageEnabled bool               `json:"tiered_storage_enabled" bson:"tiered_storage_enabled"`
